+++ packages/amplify_auth_cognito/example/ios/unit_tests/amplify_auth_cognito_tests.swift
@@ -1183,7 +1183,7 @@ class amplify_auth_cognito_tests: XCTestCase {
     func test_signOutSuccess() {
         
         class SignOutMock: AuthCognitoBridge {
-            override func onSignOut(flutterResult: @escaping FlutterResult) {
+            override func onSignOut(flutterResult: @escaping FlutterResult, request: FlutterSignOutRequest) {
                 let emptyMap: Dictionary<String, Any> = [:]
                 flutterResult(emptyMap)
             }
@@ -1202,10 +1202,43 @@ class amplify_auth_cognito_tests: XCTestCase {
         })
     }
     
+    func test_signOutGlobalSuccess() {
+        
+        class SignOutMock: AuthCognitoBridge {
+            var globalSignOutExpectation = false
+            override func onSignOut(flutterResult: @escaping FlutterResult, request: FlutterSignOutRequest) {
+                globalSignOutExpectation = request.options!.globalSignOut
+                let emptyMap: Dictionary<String, Any> = [:]
+                flutterResult(emptyMap)
+            }
+        }
+        
+        let signOutMock = SignOutMock()
+        plugin = SwiftAuthCognito.init(cognito: signOutMock)
+        
+        _data = [
+            "options": [
+                "globalSignOut": true
+            ]
+        ]
+        _args = ["data": _data]
+        let call = FlutterMethodCall(methodName: "signOut", arguments: _args)
+        plugin.handle(call, result: {(result)->Void in
+            if let res = result as? Dictionary<String, Any> {
+                XCTAssertEqual( 0, res.count )
+            } else {
+                XCTFail()
+            }
+        })
+        
+        XCTAssertTrue(signOutMock.globalSignOutExpectation)
+
+    }
+    
     func test_signOutError() {
         
         class SignOutMock: AuthCognitoBridge {
-            override func onSignOut(flutterResult: @escaping FlutterResult) {
+            override func onSignOut(flutterResult: @escaping FlutterResult, request: FlutterSignOutRequest) {
                 let authError = AuthError.invalidState("Invalid state", MockErrorConstants.invalidStateError, nil)
                 errorHandler.handleAuthError(authError: authError, flutterResult: flutterResult)
             }
@@ -1519,7 +1552,7 @@ class amplify_auth_cognito_tests: XCTestCase {
     
     func test_signInWithWebUI() {
         class SignInWithWebUIMock: AuthCognitoBridge {
-            override func onSignInWithWebUI(flutterResult: @escaping FlutterResult) {
+            override func onSignInWithWebUI(flutterResult: @escaping FlutterResult, request: FlutterSignInWithWebUIRequest) {
                 let signInRes = Result<AuthSignInResult,AuthError>.success(
                     AuthSignInResult(nextStep: AuthSignInStep.done)
                  )