+++ packages/amplify_datastore/example/integration_test/query_test/pagination_test.dart
@@ -75,6 +75,13 @@ void main() {
       expect(blogs, isEmpty);
     });
 
+    testWidgets('should default to no pagination if none is provided',
+        (WidgetTester tester) async {
+      var blogs = await Amplify.DataStore.query(Blog.classType);
+      expect(blogs.length, models.length);
+      expect(blogs, unorderedEquals(models));
+    });
+
     testWidgets('should default to a page size of 100',
         (WidgetTester tester) async {
       var blogs = await Amplify.DataStore.query(